requirements.txt
newspaper3k==0.2.8
markdown==3.5.2
beautifulsoup4==4.12.2 